moved uncertainty estimation and tools into UQ360
