+++ uq360/utils/nearest_neighbors/__init__.py
@@ -1,0 +1,1 @@
+from .base import BaseNearestNeighbors
